+++ src/packages/preset/lib/variants/breakpoints/index.ts
@@ -1,6 +1,6 @@
 /* LATER THE CLASS GENEREATES, HIGHER THE SPECIFICITY IT HAS */
 
-// breakpoints
+/* breakpoints */
 // zero: 0
 // xs: 375
 // sm: 640
@@ -8,7 +8,7 @@
 // lg: 1024
 // xl: 1280
 // mac: 1440
-// max: 1920
+// maxw: 1920
 
 import { isObject } from "../../utils";