+++ src/packages/preset/CHANGELOG.md
@@ -13,6 +13,22 @@ and this project adheres to [Semantic Versioning][semver].
 
 ---
 
+## [0.7.0] - 23-09-2023
+
+<br />
+
+### Added
+
+- #93 &nbsp;-&nbsp; Added aspect ratio
+
+### Improved
+
+- #93 &nbsp;-&nbsp; minor fixes
+
+<br /><br />
+
+---
+
 ## [0.6.0] - 06-09-2023
 
 <br>
@@ -190,6 +206,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 शुभारम्भः
 
+[0.7.0]: https://github.com/mrjadeja/vishesh/commit/f709c3c1...de281dd1
 [0.6.0]: https://github.com/mrjadeja/vishesh/commit/6e5ad8f8...f709c3c1
 [0.5.0]: https://github.com/mrjadeja/vishesh/commit/af778dea...6e5ad8f8
 [0.4.0]: https://github.com/mrjadeja/vishesh/commit/51b34689...af778dea